refactor(filter-panel): hide legacy filter panel in directory
