+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterCategory/_index.scss
@@ -4,7 +4,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../Component/mixins';
+@import '../../../Component/mixins';
 
 @import 'mixins';
