+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterCategory/_mixins.scss
@@ -4,7 +4,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../../globals/index';
+@import '../../../../globals/index';
 
 @mixin filter-category {
   /// Padding.
