+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterCategory/index.js
