+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSearch/index.js
