+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSelector/_index.scss
@@ -4,7 +4,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../Component/mixins';
+@import '../../../Component/mixins';
 
 @import 'mixins';
