+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSelector/_mixins.scss
@@ -4,7 +4,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../../globals/index';
+@import '../../../../globals/index';
 
 @mixin filter-selector {
   @include carbon--type-style($name: body-short-01);
