+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSelector/index.js
