+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSubcategory/_index.scss
@@ -3,7 +3,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../Component/mixins';
+@import '../../../Component/mixins';
 
 @import 'mixins';
