+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSubcategory/_mixins.scss
@@ -3,7 +3,7 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../../../globals/index';
+@import '../../../../globals/index';
 
 @mixin filter-subcategory {
   /// Scroller padding.
