+++ src/components/FilterPanel/LEGACY_FilterPanel/_index.scss
@@ -1,0 +1,10 @@
+////
+/// Legacy filter panel styles.
+/// @group filter-panel
+/// @copyright IBM Security 2019
+////
+
+@import './FilterCategory/index';
+@import './FilterSearch/index';
+@import './FilterSelector/index';
+@import './FilterSubcategory/index';
