+++ src/components/FilterPanel/LEGACY_FilterPanel/index.js
@@ -1,0 +1,3 @@
+import FilterPanel from './LEGACY_FilterPanel';
+
+export default FilterPanel;
