+++ src/components/FilterPanel/_index.scss
@@ -8,10 +8,7 @@
 @import '../Button/index';
 @import '../Checkbox/index';
 
-@import './FilterCategory/index';
-@import './FilterSearch/index';
-@import './FilterSelector/index';
-@import './FilterSubcategory/index';
+@import './LEGACY_FilterPanel/index';
 
 @import '../Component/mixins';
